Adding withInfo Storybook addon for better documentation.
