+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+### v1.3.4
+#### Added
+- Added `withInfo` Storybook addon to include better documentation for each component.
+
 ### v1.3.3
 #### Fixed
 - Color Contrast issue for the "danger" button style.
