+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "library-simplified-reusable-components",
-  "version": "1.3.3",
+  "version": "1.3.4",
   "description": "Reusable React components for Library Simplified interfaces",
   "repository": {
     "type": "git",
@@ -29,6 +29,7 @@
     "@babel/core": "^7.4.0",
     "@storybook/addon-a11y": "^5.0.6",
     "@storybook/addon-actions": "^5.0.6",
+    "@storybook/addon-info": "^5.0.11",
     "@storybook/addon-knobs": "^5.0.6",
     "@storybook/addon-links": "^5.0.6",
     "@storybook/addon-storysource": "^5.0.6",
@@ -52,6 +53,7 @@
     "mini-css-extract-plugin": "^0.5.0",
     "mocha": "^5.2.0",
     "node-sass": "^4.11.0",
+    "react-docgen-typescript-loader": "^3.1.0",
     "sass-lint": "^1.9.1",
     "sass-loader": "^7.1.0",
     "sinon": "^7.3.0",
