+++ stories/index.stories.tsx
@@ -1,28 +1,18 @@
 import * as React from 'react';
 
-import { storiesOf, addDecorator } from '@storybook/react';
-import { withKnobs } from '@storybook/addon-knobs';
-import { withA11y } from '@storybook/addon-a11y';
-
-import header from './header.stories';
-import panel from './panel.stories';
-import tabs from './tabs.stories';
-
-import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
-import '../src/stylesheets/app.scss';
-
-const styles = {
-  padding: '20px',
-};
-export const CenterDecorator = storyFn => <div style={styles}>{storyFn()}</div>;
-
-addDecorator(CenterDecorator);
-addDecorator(withKnobs);
-addDecorator(withA11y);
+import { storiesOf } from '@storybook/react';
 
 storiesOf('Home', module)
   .add('introduction', () => 
     <div>
       <h2>NYPL Simplified Reusable Components</h2>
-    </div>
+      <p>
+        Welcome to the documentation for components used in various SimplyE web apps.
+      </p>
+    </div>,
+    {
+      info: {
+        disable: true
+      }
+    }
   );
